AWAC - process file through KML file
