+++ AODN/WA-DOT-AWAC/README.md
@@ -3,7 +3,10 @@ Dept. Of Transport - WA - AWAC Nortek Instrument
 
 This script aims to convert wave, tide, status, current and temperature data from the Dept. of Transport into NetCDF files.
 
-The dataset was collected using a AWAC Nortek 1mhz instrument.
+This dataset was collected using a AWAC Nortek 1mhz instrument.
+
+The data is retrieved within this script using
+https://s3-ap-southeast-2.amazonaws.com/transport.wa/DOT_OCEANOGRAPHIC_SERVICES/AWAC_V2/AWAC.kml
 
 ## Using the Script
 ```bash
@@ -14,8 +17,6 @@ out the path of the new locally generated FV01 file.
 
 optional arguments:
   -h, --help            show this help message and exit
-  -i DATASET_PATH, --wave-dataset-org-path DATASET_PATH
-                        path to original wave dataset
   -o OUTPUT_PATH, --output-path OUTPUT_PATH
                         output directory of FV01 netcdf file. (Optional)
 
@@ -23,7 +24,8 @@ optional arguments:
 
 ## Location of Original Dataset
 
-The original files can be found on ```$ARCHIVE_DIR/AODN/Dept-Of-Transport_WA_WAVES```
+The dataset was previously downloaded and archived to ```$ARCHIVE_DIR/AODN/Dept-Of-Transport_WA_WAVES```
+However the most up to date dateset is found via the kml file (link above)
 
 ## Contact Support
 for support contact:
